Jira-x-lang/requirements.txt
langchain>=1.0.0,<2.0.0
langchain-core>=1.0.0,<2.0.0
langchain-anthropic>=0.1.17
python-dotenv>=1.0.1
requests>=2.32.3
Flask>=3.0.0
